+++ src/main/java/org/markdownwriterfx/preview/HtmlSourcePreview.java
@@ -5,10 +5,10 @@
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
  *
- *  o Redistributions of source code must retain the above copyright
+ *  * Redistributions of source code must retain the above copyright
  *    notice, this list of conditions and the following disclaimer.
  *
- *  o Redistributions in binary form must reproduce the above copyright
+ *  * Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
  *
@@ -28,9 +28,18 @@
 package org.markdownwriterfx.preview;
 
 import java.nio.file.Path;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashMap;
+import javafx.scene.control.IndexRange;
 import javafx.scene.control.ScrollBar;
-import javafx.scene.control.TextArea;
+import org.fxmisc.flowless.VirtualizedScrollPane;
+import org.fxmisc.richtext.StyleClassedTextArea;
+import org.fxmisc.richtext.model.StyleSpans;
+import org.fxmisc.richtext.model.StyleSpansBuilder;
 import org.markdownwriterfx.preview.MarkdownPreviewPane.Renderer;
+import org.markdownwriterfx.syntaxhighlighter.SyntaxHighlighter;
 import org.markdownwriterfx.util.Utils;
 
 /**
@@ -41,37 +50,62 @@ import org.markdownwriterfx.util.Utils;
 class HtmlSourcePreview
 	implements MarkdownPreviewPane.Preview
 {
-	private final TextArea textArea = new TextArea();
+	private final PreviewStyledTextArea textArea = new PreviewStyledTextArea();
+	private final VirtualizedScrollPane<StyleClassedTextArea> scrollPane = new VirtualizedScrollPane<>(textArea);
 	private ScrollBar vScrollBar;
 
 	HtmlSourcePreview() {
-		textArea.setEditable(false);
-		textArea.setFocusTraversable(false);
 		textArea.setWrapText(true);
+		textArea.getStylesheets().add("org/markdownwriterfx/prism.css");
 	}
 
 	@Override
 	public javafx.scene.Node getNode() {
-		return textArea;
+		return scrollPane;
 	}
 
 	@Override
 	public void update(Renderer renderer, Path path) {
-		double scrollTop = textArea.getScrollTop();
-
-		textArea.setText(renderer.getHtml());
-
-		textArea.setScrollTop(scrollTop);
+		String html = renderer.getHtml();
+		textArea.replaceText(html, computeHighlighting(html));
 	}
 
 	@Override
 	public void scrollY(double value) {
 		if (vScrollBar == null)
-			vScrollBar = Utils.findVScrollBar(textArea);
+			vScrollBar = Utils.findVScrollBar(scrollPane);
 		if (vScrollBar == null)
 			return;
 
 		double maxValue = vScrollBar.maxProperty().get();
 		vScrollBar.setValue(maxValue * value);
+	}
+
+	@Override
+	public void selectionChanged(IndexRange range) {
+	}
+
+	//---- selection highlighting ---------------------------------------------
+
+	private static final HashMap<String, Collection<String>> spanStyleCache = new HashMap<>();
+
+	private static StyleSpans<Collection<String>> computeHighlighting(String text) {
+		StyleSpansBuilder<Collection<String>> spansBuilder = new StyleSpansBuilder<>();
+		SyntaxHighlighter.highlight(text, "html", (length, style) -> {
+			spansBuilder.add(toSpanStyle(style), length);
+		});
+		return spansBuilder.create();
+	}
+
+	private static Collection<String> toSpanStyle(String style) {
+		if (style == null)
+			return Collections.emptyList();
+
+		Collection<String> spanStyle = spanStyleCache.get(style);
+		if (spanStyle == null) {
+			spanStyle = Arrays.asList(style, "token");
+			spanStyleCache.put(style, spanStyle);
+		}
+		return spanStyle;
 	}
 }